test(vscode-e2e): split createWorkspace + add rulesEngine runtime coverage
Step 2 of sub-15min CI plan. Two semantic changes:

1. Split createWorkspace.test.ts:
   - createWorkspace.fixtures.test.ts (~2-3min): drives wizard for ONLY
     the 3 runtime-fixture shapes (Standard/CustomCode/RulesEngine Stateful)
     that downstream scenario shards need. Writes manifest.
   - createWorkspace.behavior.test.ts (~11min): full 12-shape wizard
     validation + 75 assertions. Runs on its own parallel shard OFF the
     critical path.
   - createWorkspaceShared.ts: extracted helpers used by both files.

2. Add rulesEngine runtime coverage:
   - Phase 4.2 (designerActions.test.ts): NEW 3rd test case for rulesEngine
     Stateful (add trigger + action + debug + verify run). Closes the
     rulesEngine runtime-debug gap identified in
     .squad/knowledge/e2e-shape-coverage-audit.md.
   - Phase 4.3 (inlineJavascript.test.ts): parameterized via LA_E2E_SHAPE
     env var so each runtime fixture shape runs the inline-JS dep gen
     validation independently.

3. Wire scenarios[] (run-e2e.js):
   - p41-createworkspace -> p41a-fixtures (fixtures-only fast path)
   - ADD p41b-createworkspace-behavior (full 12-shape, off critical path)
   - ADD p42-customcode, p42-rulesengine
   - ADD p43-customcode, p43-rulesengine
   - runScenarioPhases() honors per-scenario env overrides (LA_E2E_SHAPE).

4. CI lint guard: run-e2e.js pre-flight check fails the build if
   *.fixtures.test.ts files contain disk-writer symbols (D-001 enforcement).
